Avoid folder deletion

+++ meteor.js
@@ -54,18 +54,20 @@ module.exports = {
     build: function(program, callback, remove){
         if (typeof remove === 'undefined') remove = true;
 
-        // remove the bundle folder
-        deleteFolderRecursive(buildPath);
+        if (remove) {
+            // remove the bundle folder
+            deleteFolderRecursive(buildPath);
+        }
 
         var command = 'meteor build '+ argPath + ' --directory';
 
         if(program.url)
              command += ' --server '+ program.url;
 
-         // if(program.settings)
-         //     command += ' --mobile-settings '+ program.settings;
+        if(program.settings)
+            command += ' --mobile-settings '+ program.settings;
 
-         // console.log('Running: '+ command);
+        console.log('Running: '+ command);
 
         execute(command, 'build the app, are you in your meteor apps folder?', callback);
     },
